# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered healthcare assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  name: string;
  email?: string | null;
  phone?: string | null;
  createdAt: string;
}

export interface CreateUserInput {
  /** @minLength 1 */
  name: string;
  email?: string | null;
  phone?: string | null;
}

export type AppointmentStatus =
  (typeof AppointmentStatus)[keyof typeof AppointmentStatus];

export const AppointmentStatus = {
  booked: "booked",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface Appointment {
  id: string;
  userId: string;
  doctorName: string;
  specialty: string;
  date: string;
  time: string;
  reason?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}

export interface CreateAppointmentInput {
  userId: string;
  /** @minLength 1 */
  doctorName: string;
  /** @minLength 1 */
  specialty: string;
  date: string;
  time: string;
  reason?: string | null;
}

export type UpdateAppointmentInputStatus =
  (typeof UpdateAppointmentInputStatus)[keyof typeof UpdateAppointmentInputStatus];

export const UpdateAppointmentInputStatus = {
  booked: "booked",
  cancelled: "cancelled",
  completed: "completed",
} as const;

export interface UpdateAppointmentInput {
  doctorName?: string;
  specialty?: string;
  date?: string;
  time?: string;
  reason?: string | null;
  status?: UpdateAppointmentInputStatus;
}

export interface ChatInput {
  userId: string;
  /** @minLength 1 */
  message: string;
}

export type ChatResponseSuggestion =
  | (typeof ChatResponseSuggestion)[keyof typeof ChatResponseSuggestion]
  | null;

export const ChatResponseSuggestion = {
  book_appointment: "book_appointment",
  see_doctor: "see_doctor",
  none: "none",
} as const;

export interface ChatResponse {
  id: string;
  userId: string;
  message: string;
  response: string;
  suggestion?: ChatResponseSuggestion;
  createdAt: string;
}

export interface Conversation {
  id: string;
  userId: string;
  message: string;
  response: string;
  createdAt: string;
}

export interface DashboardSummary {
  user: User;
  upcomingAppointments: Appointment[];
  pastAppointments: Appointment[];
  totalAppointments: number;
  totalConversations: number;
  recentConversations: Conversation[];
}