# lib/api-zod/src/generated/types/appointment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered healthcare assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentStatus } from "./appointmentStatus";

export interface Appointment {
  id: string;
  userId: string;
  doctorName: string;
  specialty: string;
  date: string;
  time: string;
  reason?: string | null;
  status: AppointmentStatus;
  createdAt: string;
}