# lib/api-zod/src/generated/types/chatInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered healthcare assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ChatInput {
  userId: string;
  /** @minLength 1 */
  message: string;
}