# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered healthcare assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatResponseSuggestion } from "./chatResponseSuggestion";

export interface ChatResponse {
  id: string;
  userId: string;
  message: string;
  response: string;
  suggestion?: ChatResponseSuggestion;
  createdAt: string;
}