# lib/api-zod/src/generated/types/createAppointmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered healthcare assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAppointmentInput {
  userId: string;
  /** @minLength 1 */
  doctorName: string;
  /** @minLength 1 */
  specialty: string;
  date: string;
  time: string;
  reason?: string | null;
}