# lib/api-zod/src/generated/types/createUserInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered healthcare assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateUserInput {
  /** @minLength 1 */
  name: string;
  email?: string | null;
  phone?: string | null;
}