# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered healthcare assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { Appointment } from "./appointment";
import type { Conversation } from "./conversation";
import type { User } from "./user";

export interface DashboardSummary {
  user: User;
  upcomingAppointments: Appointment[];
  pastAppointments: Appointment[];
  totalAppointments: number;
  totalConversations: number;
  recentConversations: Conversation[];
}