# lib/api-zod/src/generated/types/updateAppointmentInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-powered healthcare assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateAppointmentInputStatus } from "./updateAppointmentInputStatus";

export interface UpdateAppointmentInput {
  doctorName?: string;
  specialty?: string;
  date?: string;
  time?: string;
  reason?: string | null;
  status?: UpdateAppointmentInputStatus;
}